mvp1_classification/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==2.0.1
torch==2.4.0
transformers==4.44.2
timm==0.9.16
